release 10.0.39427

+++ core/tools/md5.h
@@ -7,7 +7,7 @@ namespace core
 {
     namespace tools
     {
-        std::string md5(const void* _data, int32_t _size);
+        std::string md5(const void* _data, size_t _size);
 
         uint8_t hextobin(const char * str, uint8_t * bytes, size_t blen);
     }
